Refactor collation check and fix into separate class
